Replace some hardcoded refs to playground.wordpress.net web app (#42)
## Motivation for the change, related issues

We are deploying a private Playground instance and want it to reference
itself rather than the original public site. In addition, these changes
will help folks who also want to self-host.

## Implementation details

In general, this PR replaces references to playground.wordpress.net with
relative URL references where those references might interfere with
Playground platform advances deployed to a private Playground instance.

There are a number of references to
`https://playground.wordpress.net/plugin-proxy.php`, and I am leaving
those alone for now. `plugin-proxy.php` requires a GITHUB_TOKEN
environment var, which requires further configuration, and this would be
particularly annoying for folks running local Playground servers. For
now, let's just keep using the public hosted version of that script and
adjust later if required.

In other cases, like that of the reference to
https://playground.wordpress.net/logger.php, I am leaving the references
as-is until there is a benefit to changing them.

## Testing Instructions (or ideally a Blueprint)

- CI