Website: Blueprints as PWAs with a dynamic manifest.json file (#67)
A static manifest.json contains `"start_url": "/"`, which means every
subsequent visit will open a plain vanilla Playground.

This PR rewires the Playground website to use a dynamic, PHP-generated
manifest file. It renders the same manifest as before, except the
hostname is not hardcoded and the `"start_url"` field matches the URL
used to visit Playground.

This means you can go to a Playground site with a specific Blueprint
applied, give it an `?app_name=rss_reader`, and install that customized
version of Playground as an offline PWA on your device.

 ## Testing instructions

Build Playground, run it via a PHP dev server on a https domain (I've
used ngrok), start it with a Blueprint and a custom `?app_name`
parameter, and confirm you can install it as a named PWA that runs that
Blueprint on boot.

cc @bgrgicak @brandonpayton  @akirk