Fix failing WP update workflows (#2376)
## Motivation for the change, related issues

Our automation that updates WordPress builds is currently broken. The
workflows try to run Playground CLI with bun, and that is not currently
working (due to broken worker URL resolution at runtime).

## Implementation details

This PR adjusts the build script and the workflows to use Node.js
instead of Bun.

It also fixes a couple other small bugs:
1. The Playground CLI Nx targets `unbuilt-asyncify` and `unbuilt-jspi`
were mistakenly passing the `--watch` argument to node. My intent with
those targets was for them to be easy ways to run Playground CLI once.
2. The WP build script were exiting with an error if the beta version
was specified but a beta version wasn't available. Sometimes beta
versions aren't available, so now we exit normally in that case.

## Testing Instructions (or ideally a Blueprint)

- CI
- Deploy
- Test workflows on trunk

These workflows are already broken, so I'm just going merge and see if
they work in production. If there are further issues, I'll open another
PR and do the slower kind of testing that involves commenting out
only-on-main-branch guards and temporarily adding exceptions to branch
protections.