Fix ESLint 9 linting in VSCode (#2417)
## Motivation for the change, related issues

Upgrading to ESLint 9 broke real-time linter feedback in VSCode because
ESLint 9 defaults to a new "flat config" format which we are not using.

Fixes #2416

## Implementation details

This PR restores real-time ESLint feedback by adjusting VSCode workspace
settings to use the legacy config format.

This PR also adds an explicit "useFlagConfig: false" setting to Nx lint
targets. Hopefully, this will prevent NX-based linting breakage when we
upgrade to new major Nx version.

## Testing Instructions (or ideally a Blueprint)

- CI
- Open a `.ts` file in VSCode, add a `console.log()` statement, and
observe the the console statement is included in the Problems tab.

Co-authored-by: Adam Zieliński <adam@adamziel.com>