[CI] Do not block CI tests on linter (#2619)
## Motivation for the change, related issues

I often run into linting issues that prevent me from knowing whether all
the tests pass on my PR. With this PR, tests will run without waiting
for the linter. If all the tests puss but the linting doesn't, it's a
good signal we can solve the formatting issues and merge.