Fixing old documentation regarding PHP code examples (#2721)
## Motivation for the change, related issues
Context: https://github.com/WordPress/wordpress-playground/issues/2686

Documentation should be updated to reflect the new change, where the
root folder has been moved from `/` to `/wordpress` and not a relative
path. So all the old `require_once 'wordpress/wp-load'` should now be
documented as a `require_once '/wordpress/wp-load'` to be sure not
having a warning in the console.

## Implementation details
Fixed some comments and documentation

## Testing Instructions (or ideally a Blueprint)
Nothing to test in particular, a quick check at the code is enough.