[Website] Remove "preview WordPress core branch" feature (#2894)
## Motivation for the change, related issues

It doesn't seem like `wordpress-develop` is storing artifacts for the
trunk branch:

https://github.com/WordPress/wordpress-develop/actions/runs/19318833074

This PR removes the part of the UI that suggests you can preview any
built WordPress core branch. For PRs, there's still the PR number input,
and for trunk we have a `nightly` WordPress version available in the
picker that's a close approximation.

## Testing Instructions (or ideally a Blueprint)

Confirm the "Import core PR" form doesn't offer to accept a branch name.