[Blueprints] Don't treat extra output during plugin activation as a failure (#2904)
Some WordPress plugins emit output during activation. Before this PR,
Playground treated that as an activation error. But that's not right,
the plugin may still get activated. This PR checks in with WordPress
whether or not the plugin is activated after we're tried to activate it.
If it is, that's our success.

## Testing Instructions (or ideally a Blueprint)

CI, this PR ships a new test.